Nope. Still doesn't work.
grr. The total energy seems to be fluctuating. May have something to do
with calculation of either potential or kinetic energy: if they are
taken as being off by a phase shift, they return a flat total energy.
Also, added potential energy between pieces, but it's off by an order of
magnitude. Also, need to look for energy not calculated.